+++ src/Ports.php
@@ -13,11 +13,8 @@ use Innmind\Immutable\Attempt;
 
 final class Ports
 {
-    private Config $config;
-
-    private function __construct(Config $config)
+    private function __construct(private Config $config)
     {
-        $this->config = $config;
     }
 
     /**